+++ baikai/test/Main.hs
@@ -31,6 +31,7 @@ import GHC.Generics (C1, D1, Rep, S1, Selector (selName), (:*:))
 import GenModelsSpec qualified
 import HelpersSpec qualified
 import InteractiveSpec qualified
+import PricingPolicySpec qualified
 import PublicSurfaceSpec qualified
 import StreamSpec qualified
 import StreamWorkerSpec qualified
@@ -118,6 +119,7 @@ main = do
         GenModelsSpec.tests,
         HelpersSpec.tests,
         InteractiveSpec.tests,
+        PricingPolicySpec.tests,
         PublicSurfaceSpec.tests,
         StreamSpec.tests,
         StreamWorkerSpec.tests,